Gate the plugin-command dispatcher behind the plugins feature
handle_plugin_command (a ~1300-line match) and all per-command handlers
reachable only from it were compiled unconditionally under runtime, even
though without plugins PluginManager::process_commands() always returns
empty and nothing can drive them.

Extract the SHARED widget/floating-panel runtime (the methods reached from
non-plugin input/mouse/click/lifecycle/render code) into a new un-gated
module app/widget_runtime.rs (~2250 lines), leaving plugin_dispatch.rs and
plugin_commands.rs as purely plugin-only. Gate those two modules and the
three command-drain call sites (render.rs x2, async_messages/async_dispatch)
behind #[cfg(feature = "plugins")]. Plugin-only helpers orphaned by the
gating (in buffer_groups/composite_buffer_actions/editor_accessors and a few
mod.rs items) are gated to match.

Gate the plugin-dependent integration tests (the e2e modules that drive
PluginCommands, plus regression_hidden_cursor_panic) behind plugins so the
CI 'cargo check --no-default-features --features runtime --all-targets' gate
stays green; they still build and run under --all-features.

Both CI compile gates are green and the default build behavior is unchanged.
