+++ crates/fresh-editor/src/app/plugin_commands.rs
@@ -893,32 +893,6 @@ impl Editor {
         }
     }
 
-    /// Handle CloseSplit command
-    pub(super) fn handle_close_split(&mut self, split_id: SplitId) {
-        // Plugin sends arbitrary SplitId — convert to LeafId at the boundary
-        let leaf_id = LeafId(split_id);
-        match self
-            .windows
-            .get_mut(&self.active_window)
-            .and_then(|w| w.split_manager_mut())
-            .expect("active window must have a populated split layout")
-            .close_split(leaf_id)
-        {
-            Ok(()) => {
-                // Clean up the view state for the closed split
-                self.windows
-                    .get_mut(&self.active_window)
-                    .and_then(|w| w.split_view_states_mut())
-                    .expect("active window must have a populated split layout")
-                    .remove(&leaf_id);
-                tracing::info!("Closed split {:?}", split_id);
-            }
-            Err(e) => {
-                tracing::warn!("Failed to close split {:?}: {}", split_id, e);
-            }
-        }
-    }
-
     /// Handle SetSplitRatio command
     pub(super) fn handle_set_split_ratio(&mut self, split_id: SplitId, ratio: f32) {
         // Plugin sends arbitrary SplitId — convert to ContainerId at the boundary
@@ -1844,19 +1818,6 @@ impl Editor {
         }
     }
 
-    /// Handle RefreshLines command
-    pub(super) fn handle_refresh_lines(&mut self, buffer_id: BufferId) {
-        // Clear seen_byte_ranges for this buffer so all visible lines will be re-processed
-        // on the next render. This is useful when a plugin is enabled and needs to
-        // process lines that were already marked as seen.
-        self.active_window_mut().seen_byte_ranges.remove(&buffer_id);
-        // Request a render so the lines_changed hook fires
-        #[cfg(feature = "plugins")]
-        {
-            self.plugin_render_requested = true;
-        }
-    }
-
     /// Handle SetLineIndicator command
     pub(super) fn handle_set_line_indicator(
         &mut self,
@@ -2417,201 +2378,6 @@ impl Editor {
         self.grammar_reload_pending = true;
         self.pending_grammar_callbacks.push(callback_id);
     }
-
-    /// Flush pending grammars: spawn a background rebuild if any ReloadGrammars
-    /// commands were received during this command batch.
-    ///
-    /// Called after processing all plugin commands in a batch, so that multiple
-    /// RegisterGrammar+ReloadGrammars pairs result in only one rebuild.
-    /// The rebuild happens on a background thread; when complete, a
-    /// `GrammarRegistryBuilt` message swaps in the new registry.
-    ///
-    /// On the first call, this triggers the deferred full grammar build
-    /// (user grammars + language packs + any plugin grammars accumulated so far).
-    pub(super) fn flush_pending_grammars(&mut self) {
-        // On the first call, start the deferred full grammar build.
-        // This includes any plugin grammars that were registered during init,
-        // so we get everything in a single builder.build() pass.
-        if self.needs_full_grammar_build {
-            self.needs_full_grammar_build = false;
-            self.grammar_reload_pending = false;
-
-            // Drain all pending grammars to include in the initial build
-            let additional: Vec<_> = self
-                .pending_grammars
-                .drain(..)
-                .map(|g| crate::primitives::grammar::GrammarSpec {
-                    language: g.language.clone(),
-                    path: std::path::PathBuf::from(g.grammar_path),
-                    extensions: g.extensions.clone(),
-                })
-                .collect();
-
-            // Update config.languages with the extensions so detect_language() works
-            for crate::primitives::grammar::GrammarSpec {
-                language,
-                extensions,
-                ..
-            } in &additional
-            {
-                let lang_config = self
-                    .config_mut()
-                    .languages
-                    .entry(language.clone())
-                    .or_default();
-                for ext in extensions {
-                    if !lang_config.extensions.contains(ext) {
-                        lang_config.extensions.push(ext.clone());
-                    }
-                }
-            }
-
-            let callback_ids: Vec<_> = self.pending_grammar_callbacks.drain(..).collect();
-            self.start_background_grammar_build(additional, callback_ids);
-            return;
-        }
-
-        if !self.grammar_reload_pending {
-            return;
-        }
-        self.grammar_reload_pending = false;
-
-        // If a background build is already in progress, it will call
-        // flush_pending_grammars() again when it completes — so just
-        // re-arm the flag and return.
-        if self.grammar_build_in_progress {
-            self.grammar_reload_pending = true;
-            tracing::debug!("Grammar build in progress, deferring flush");
-            return;
-        }
-
-        use std::path::PathBuf;
-
-        if self.pending_grammars.is_empty() {
-            tracing::debug!("Grammar reload requested but no pending grammars");
-            return;
-        }
-
-        // Deduplicate: skip grammars whose extensions are all already mapped
-        // in the current registry (meaning the grammar was already loaded by
-        // for_editor or a previous build).
-        let pending_before = self.pending_grammars.len();
-        self.pending_grammars.retain(|g| {
-            // Check if ALL extensions for this grammar are already mapped
-            let all_mapped = !g.extensions.is_empty()
-                && g.extensions
-                    .iter()
-                    .all(|ext| self.grammar_registry.find_by_extension(ext).is_some());
-            if all_mapped {
-                tracing::debug!(
-                    "Skipping already-loaded grammar '{}' (extensions {:?} already mapped)",
-                    g.language,
-                    g.extensions
-                );
-                false
-            } else {
-                true
-            }
-        });
-        if pending_before != self.pending_grammars.len() {
-            tracing::info!(
-                "Deduplicated pending grammars: {} -> {}",
-                pending_before,
-                self.pending_grammars.len()
-            );
-        }
-
-        if self.pending_grammars.is_empty() {
-            tracing::info!(
-                "All pending grammars already loaded, resolving callbacks without rebuild"
-            );
-            // Resolve callbacks immediately — no rebuild needed
-            #[cfg(feature = "plugins")]
-            for cb_id in self.pending_grammar_callbacks.drain(..) {
-                self.plugin_manager
-                    .read()
-                    .unwrap()
-                    .resolve_callback(cb_id, "null".to_string());
-            }
-            #[cfg(not(feature = "plugins"))]
-            self.pending_grammar_callbacks.clear();
-            return;
-        }
-
-        tracing::info!(
-            "Flushing {} pending grammars via background rebuild",
-            self.pending_grammars.len()
-        );
-
-        // Collect pending grammars
-        let additional: Vec<crate::primitives::grammar::GrammarSpec> = self
-            .pending_grammars
-            .drain(..)
-            .map(|g| crate::primitives::grammar::GrammarSpec {
-                language: g.language.clone(),
-                path: PathBuf::from(g.grammar_path),
-                extensions: g.extensions.clone(),
-            })
-            .collect();
-
-        // Update config.languages with the extensions so detect_language() works
-        for crate::primitives::grammar::GrammarSpec {
-            language,
-            extensions,
-            ..
-        } in &additional
-        {
-            let lang_config = self
-                .config_mut()
-                .languages
-                .entry(language.clone())
-                .or_default();
-            for ext in extensions {
-                if !lang_config.extensions.contains(ext) {
-                    lang_config.extensions.push(ext.clone());
-                }
-            }
-        }
-
-        // Collect pending callback IDs to resolve when build completes
-        let callback_ids: Vec<_> = self.pending_grammar_callbacks.drain(..).collect();
-
-        // Spawn background rebuild
-        let base_registry = std::sync::Arc::clone(&self.grammar_registry);
-        if let Some(bridge) = &self.async_bridge {
-            let sender = bridge.sender();
-            self.grammar_build_in_progress = true;
-            std::thread::Builder::new()
-                .name("grammar-rebuild".to_string())
-                .spawn(move || {
-                    use crate::primitives::grammar::GrammarRegistry;
-                    match GrammarRegistry::with_additional_grammars(&base_registry, &additional) {
-                        Some(new_registry) => {
-                            // Ok to ignore: receiver may be gone if app is shutting down.
-                            drop(sender.send(
-                                crate::services::async_bridge::AsyncMessage::GrammarRegistryBuilt {
-                                    registry: std::sync::Arc::new(new_registry),
-                                    callback_ids,
-                                },
-                            ));
-                        }
-                        None => {
-                            tracing::error!("Failed to rebuild grammar registry in background");
-                            // Still send the message so callbacks get resolved (even on failure)
-                            drop(sender.send(
-                                crate::services::async_bridge::AsyncMessage::GrammarRegistryBuilt {
-                                    registry: base_registry,
-                                    callback_ids,
-                                },
-                            ));
-                        }
-                    }
-                })
-                .ok();
-        }
-    }
-
-    // ==================== Project Grep ====================
 
     /// Handle GrepProject command: walk files, search buffers/disk, collect matches
     pub(super) fn handle_grep_project(
@@ -3407,7 +3173,7 @@ impl Editor {
         if area.width == 0 || area.height == 0 {
             return;
         }
-        let animation_kind = translate_plugin_animation_kind(kind);
+        let animation_kind = super::widget_runtime::translate_plugin_animation_kind(kind);
         self.active_window_mut().animations.start_with_id(
             crate::view::animation::AnimationId::from_raw(id),
             area,
@@ -3432,7 +3198,7 @@ impl Editor {
         }
         match self.virtual_buffer_screen_rect(buffer_id) {
             Some(area) => {
-                let animation_kind = translate_plugin_animation_kind(kind);
+                let animation_kind = super::widget_runtime::translate_plugin_animation_kind(kind);
                 self.active_window_mut().animations.start_with_id(
                     crate::view::animation::AnimationId::from_raw(id),
                     area,
@@ -3447,72 +3213,6 @@ impl Editor {
                 self.pending_vb_animations.push((id, buffer_id, kind));
             }
         }
-    }
-
-    /// Retry deferred virtual-buffer animations now that split_areas has
-    /// been recomputed. Called from render() after layout but before
-    /// animations.apply_all so the first frame of the effect lands in
-    /// the same render pass.
-    pub(crate) fn drain_pending_vb_animations(&mut self) {
-        if self.pending_vb_animations.is_empty() {
-            return;
-        }
-        let pending = std::mem::take(&mut self.pending_vb_animations);
-        for (id, buffer_id, kind) in pending {
-            match self.virtual_buffer_screen_rect(buffer_id) {
-                Some(area) => {
-                    let animation_kind = translate_plugin_animation_kind(kind);
-                    self.active_window_mut().animations.start_with_id(
-                        crate::view::animation::AnimationId::from_raw(id),
-                        area,
-                        animation_kind,
-                    );
-                }
-                None => {
-                    // Still not visible; keep pending for next frame.
-                    self.pending_vb_animations.push((id, buffer_id, kind));
-                }
-            }
-        }
-    }
-
-    /// Look up the on-screen Rect currently occupied by `buffer_id`, if any.
-    /// Reads from the cached split layout captured in the last render pass.
-    pub(crate) fn virtual_buffer_screen_rect(
-        &self,
-        buffer_id: BufferId,
-    ) -> Option<ratatui::layout::Rect> {
-        self.active_layout()
-            .split_areas
-            .iter()
-            .find(|(_, bid, _, _, _, _)| *bid == buffer_id)
-            .map(|(_, _, content_rect, _, _, _)| *content_rect)
-    }
-}
-
-/// Translate the plugin-facing animation description to the internal
-/// `AnimationKind` the runner consumes.
-fn translate_plugin_animation_kind(
-    kind: fresh_core::api::PluginAnimationKind,
-) -> crate::view::animation::AnimationKind {
-    use crate::view::animation::{AnimationKind, Edge};
-    use fresh_core::api::{PluginAnimationEdge, PluginAnimationKind};
-    use std::time::Duration;
-    match kind {
-        PluginAnimationKind::SlideIn {
-            from,
-            duration_ms,
-            delay_ms,
-        } => AnimationKind::SlideIn {
-            from: match from {
-                PluginAnimationEdge::Top => Edge::Top,
-                PluginAnimationEdge::Bottom => Edge::Bottom,
-                PluginAnimationEdge::Left => Edge::Left,
-                PluginAnimationEdge::Right => Edge::Right,
-            },
-            duration: Duration::from_millis(duration_ms as u64),
-            delay: Duration::from_millis(delay_ms as u64),
-        },
     }
 }
